refactor: move ip whitelist query to forestadmin-client
