+++ packages/agent/test/utils/forest-http-api.test.ts
@@ -19,74 +19,6 @@ describe('ForestHttpApi', () => {
     jest.restoreAllMocks();
   });
 
-  describe('getIpWhitelist', () => {
-    describe('the syntax of the rules', () => {
-      it.each([
-        { ipMinimum: '10.10.1.2', ipMaximum: '10.10.1.3' },
-        { ip: '10.10.1.2' },
-        { range: '10.10.1.2/24' },
-      ])(`%s should be supported`, async rule => {
-        const ipRules = [{ type: 1, ...rule }];
-        const isFeatureEnabled = true;
-
-        superagentMock.set.mockResolvedValue({
-          body: {
-            data: {
-              attributes: {
-                use_ip_whitelist: isFeatureEnabled,
-                rules: ipRules,
-              },
-            },
-          },
-        });
-
-        const result = await ForestHttpApi.getIpWhitelistConfiguration(options);
-
-        expect(result).toStrictEqual({ isFeatureEnabled, ipRules });
-      });
-    });
-
-    test('should fetch the correct end point with the env secret', async () => {
-      superagentMock.set.mockResolvedValue({
-        body: {
-          data: {
-            attributes: {
-              use_ip_whitelist: true,
-              rules: [],
-            },
-          },
-        },
-      });
-
-      await ForestHttpApi.getIpWhitelistConfiguration(options);
-
-      expect(superagentMock.set).toHaveBeenCalledWith('forest-secret-key', 'myEnvSecret');
-      expect(superagentMock.get).toHaveBeenCalledWith(
-        'https://api.url/liana/v1/ip-whitelist-rules',
-      );
-    });
-
-    test('should return the ip rules and the isFeatureEnabled attributes', async () => {
-      const ipRules = [{ type: 1, ip: '10.20.15.10' }];
-      const isFeatureEnabled = true;
-
-      superagentMock.set.mockResolvedValue({
-        body: {
-          data: {
-            attributes: {
-              use_ip_whitelist: isFeatureEnabled,
-              rules: ipRules,
-            },
-          },
-        },
-      });
-
-      const result = await ForestHttpApi.getIpWhitelistConfiguration(options);
-
-      expect(result).toStrictEqual({ isFeatureEnabled, ipRules });
-    });
-  });
-
   describe('getOpenIdIssuerMetadata', () => {
     test('should fetch the correct end point with the env secret', async () => {
       superagentMock.set.mockResolvedValue({
@@ -210,7 +142,7 @@ describe('ForestHttpApi', () => {
         async status => {
           superagentMock.set.mockRejectedValue({ response: { status } });
 
-          await expect(ForestHttpApi.getIpWhitelistConfiguration(options)).rejects.toThrow(
+          await expect(ForestHttpApi.getOpenIdIssuerMetadata(options)).rejects.toThrow(
             /Are you online/,
           );
         },
