+++ packages/forestadmin-client/src/index.ts
@@ -21,6 +21,7 @@ export {
   ChartHandlerInterface,
   ContextVariablesInstantiatorInterface,
 } from './types';
+export { IpWhitelistConfiguration } from './ip-whitelist/types';
 export { CollectionActionEvent } from './permissions/types';
 
 export default function createForestAdminClient(
