+++ packages/forestadmin-client/src/ip-whitelist/types.ts
@@ -1,0 +1,10 @@
+export type IpWhitelistConfiguration = {
+  isFeatureEnabled: boolean;
+  ipRules: Array<{
+    type: number;
+    ipMinimum?: string;
+    ipMaximum?: string;
+    ip?: string;
+    range?: string;
+  }>;
+};
